+++ tomcat/src/main/java/org/apache/catalina/controller/RegisterController.java
@@ -2,7 +2,7 @@ package org.apache.catalina.controller;
 
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
-import org.apache.catalina.HeaderName;
+import org.apache.catalina.http.HeaderName;
 import org.apache.catalina.http.StatusCode;
 import org.apache.catalina.request.HttpRequest;
 import org.apache.catalina.response.HttpResponse;
