+++ tomcat/src/main/java/org/apache/catalina/exception/UncheckedServletException.java
@@ -1,8 +1,8 @@
-package com.techcourse.exception;
+package org.apache.catalina.exception;
 
 public class UncheckedServletException extends RuntimeException {
 
     public UncheckedServletException(Exception e) {
-        super(e);
+        super("UncheckedServletException occured", e);
     }
 }
