+++ tomcat/src/main/java/org/apache/catalina/request/RequestHeader.java
@@ -3,7 +3,7 @@ package org.apache.catalina.request;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import org.apache.catalina.HeaderName;
+import org.apache.catalina.http.HeaderName;
 import org.apache.catalina.coockie.HttpCookie;
 
 public class RequestHeader {
